% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_viral_production.R
\name{analyze_viral_production}
\alias{analyze_viral_production}
\title{Title}
\usage{
analyze_viral_production(
  vp_results = data.frame(),
  data = data.frame(),
  original_abundances = data.frame(),
  burst_sizes = c(),
  bacterial_secondary_production = NULL,
  nutrient_content_bacteria = list(),
  nutrient_content_viruses = list(),
  write_csv = TRUE
)
}
\arguments{
\item{vp_results}{test}

\item{data}{test}

\item{original_abundances}{test}

\item{burst_sizes}{test}

\item{bacterial_secondary_production}{test}

\item{nutrient_content_bacteria}{test}

\item{nutrient_content_viruses}{test}

\item{write_csv}{test}
}
\value{
test
}
\description{
test
}
\examples{
\dontrun{
test

}
}
