% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_dataframe_functions.R
\name{vp_average_replicate_dataframe}
\alias{vp_average_replicate_dataframe}
\title{Average replicate data frame}
\usage{
vp_average_replicate_dataframe(data, add_timepoints = TRUE)
}
\arguments{
\item{data}{Dataframe with the output of the flow cytometer. Data frame acquired by reading in the data of csv file.}

\item{add_timepoints}{If \code{TRUE}, different time ranges will be added with \code{vp_add_timepoints()}. If \code{FALSE}, no addition of time ranges to data frame. (Default = \code{TRUE})}
}
\value{
When \code{add_timepoints = FALSE}, a data frame with 11 columns. When \code{add_timepoints = TRUE}, a data frame with 13 columns is returned since the time ranges are added.
}
\description{
Determining the bacterial and viral counts by taking the average over the replicates.
}
\examples{
\dontrun{
data_NJ2020 <- read.csv(system.file('extdata', 'NJ2020_subset.csv', package = "viralprod"))
vp_average_replicate_dataframe(data_NJ2020)
vp_average_replicate_dataframe(data_NJ2020, add_timepoints = F)
}
}
