% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_viral_production.R
\name{calculate_viral_production}
\alias{calculate_viral_production}
\title{Wrapper function that calculates viral production}
\usage{
calculate_viral_production(
  data = data.frame(),
  methods = c(1:12),
  SR_calc = TRUE,
  BP_endpoint = TRUE,
  write_csv = TRUE,
  output_dir = ""
)
}
\arguments{
\item{data}{Data frame with the output of the flow cytometry.}

\item{methods}{Integer vector with the indexes of \code{list_of_methods}. Indexes determine which methods will be used for viral production calculation.}

\item{SR_calc}{If \code{TRUE}, separate replicate treatment results will be performed and saved as an separate data frame.
Set to \code{FALSE}, if separate replicate treatment results are not wanted. (Default = \code{TRUE})}

\item{BP_endpoint}{If \code{TRUE}, the bacterial endpoint will be taken into account and only those results will be saved in a new
data frame. If not wanted, set to \code{FALSE}. (Default = \code{TRUE})}

\item{write_csv}{If \code{TRUE}, the output data frames will be saved as csv files in a folder specified by \code{output_dir}.
If no csv files are wanted, set to \code{FALSE}. (Default = \code{TRUE})}

\item{output_dir}{String that refers to the location of folder to save the data frames as csv files.}
}
\value{
Depending on setting of parameters, different data frames with the viral production calculation will be available in the global environment.
}
\description{
Wrapper function that performs viral production calculation at once. Given the output of the flow cytometry step, viral production
is calculated for all the methods of linear regression and VIPCAL. The output data frames are written as csv files,
four different csv files are saved:
\enumerate{
\item \code{vp_results_ALL.csv}: contains the results of all the samples for each time range.
\item \code{vp_results_T24.csv}: contains the results of all the samples for the end of the assay (T0_T24).
\item \code{vp_results_SR.csv}: contains the results of the separate replicate treatment of all the samples for each time range.
\item \code{vp_results_BP.csv}: contains the results of all the samples but takes the bacterial endpoint into account, bacterial endpoint decides time range of the results.
}

More details about each of the methods used:
\itemize{
\item Linear regression variants: \link{vp_methods_linear}
\item VIPCAL variants: \link{vp_methods_VIPCAL}
}
}
\examples{
\dontrun{
data_NJ2020_all <- read.csv(system.file('extdata', 
'NJ2020_Station_2_and_6_all_populations.csv', package = "viralprod"))

calculate_viral_production(data_NJ2020_all, 
output_dir = paste0(system.file(“extdata”, package = “viralprod”), 
“/vp_results_NJ2020”))

calculate_viral_production(data_NJ2020_all, write_csv = F)

calculate_viral_production(data_NJ2020_all, write_csv = F, SR_calc = F, BP_endpoint = F)

calculate_viral_production(data_NJ2020_all, write_csv = F, methods = c(2,3,8,12))
}
}
