% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_to_end_viral_production.R
\name{end_to_end_viral_production}
\alias{end_to_end_viral_production}
\title{End-to-End viral production analysis}
\usage{
end_to_end_viral_production(
  data = data.frame(),
  original_abundances = data.frame(),
  methods = c(1:12),
  SR_calc = TRUE,
  BP_endpoint = TRUE,
  burst_sizes = c(),
  bacterial_secondary_production = NULL,
  nutrient_content_bacteria = list(),
  nutrient_content_viruses = list(),
  write_output = TRUE,
  output_dir = ""
)
}
\arguments{
\item{data}{Data frame with the output of the flow cytometry.}

\item{original_abundances}{Data frame with the abundances of bacterial and virus population in the original sample.}

\item{methods}{Integer vector with the indexes of \code{list_of_methods}. Indexes determine which methods will be performed within the viral production calculation.}

\item{SR_calc}{If \code{TRUE}, separate replicate treatment results will be stored and saved as an separate data frame.
Set to \code{FALSE}, if separate replicate treatment results are not wanted. (Default = \code{TRUE})}

\item{BP_endpoint}{If \code{TRUE}, the bacterial endpoint will be taken into account and only those results will be saved in a new
data frame. If not wanted, set to \code{FALSE}. (Default = \code{TRUE})}

\item{burst_sizes}{Vector with three different burst sizes. The burst size refers to the number of new viral particles released from an infected bacterial cell.}

\item{bacterial_secondary_production}{Value for the bacterial secondary production, how much new bacterial biomass is produced as a result of bacterial growth and reproduction.}

\item{nutrient_content_bacteria}{List with the amount of organic carbon, nitrogen and phosphor released by a bacteria, preferred a aquatic, North Sea bacteria.}

\item{nutrient_content_viruses}{List with the amount of organic carbon, nitrogen and phosphor released by a marine virus (bacteriophage).}

\item{write_output}{If \code{TRUE}, the output data frames will be saved as csv files in the same folder of the viral production calculation.
If no csv files are wanted, set to \code{FALSE}. (Default = \code{TRUE})}

\item{output_dir}{String that refers to the location of folder to save the data frames as csv files.}
}
\value{
All outputs will be saved in the folder specified by \code{output_dir}.
}
\description{
Wrapper function that includes everything from the \code{viralprod} package. Calculation, analyzing and
visualization of viral production based on the input data from the flow cytometry and the abundances
of bacterial and virus populations in the original sample. All outputs will be saved together in the
folder specified by the argument \code{output_dir}.

More details about the previous performed steps can be found on:
\enumerate{
\item Calculate viral production step: \link{calculate_viral_production}
\item Analyze viral production step: \link{analyze_viral_production}
\item Visualize viral production data: \link{visualize_viral_production}
}
}
\examples{
\dontrun{
# Setup
data_NJ2020_all <- read.csv(system.file('extdata', 
'NJ2020_Station_2_and_6_all_populations.csv', package = "viralprod"))

original_abundances_NJ2020 <- read.csv(system.file('extdata',
'NJ2020_original_abundances.csv', package = "viralprod"))

# Perform
end_to_end_viral_production <- function(
data_NJ2020_all,
original_abundances_NJ2020,
output_dir = paste0(system.file(“extdata”, package = “viralprod”), 
“/vp_results_NJ2020_all”))
}
}
