% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_class_viral_production.R
\name{S3_class_viral_production}
\alias{S3_class_viral_production}
\alias{new_viralprod_class}
\alias{new_viralprod_class_2}
\title{Create S3 class for input data frames}
\usage{
new_viralprod_class(data = data.frame())

new_viralprod_class_2(original_abundances = data.frame())
}
\arguments{
\item{data}{Data frame with the output of the flow cytometry.}

\item{original_abundances}{Data frame with the abundances of bacterial and virus population in the original sample.}
}
\value{
If data frame meets the requirements, a new class is added. Otherwise, an error will be thrown.
}
\description{
The \code{viralprod} package is dependent on two data frames, the output data frame of the flow cytometry step and a data
frame with the original abundances in the sample. To make the R package more organized, maintainable and user-friendly,
\verb{S3 object-oriented programming} is introduced. Based on the requirements of the data frames, a separate S3 class is defined.
The functions to calculate, analyze and visualize the viral production from the viral reduction assay will only work on these specific
S3 classes, otherwise an error will be thrown.

\code{new_viralprod_class} checks the output data frame of the flow cytometry step. If the requirements are met, \code{viralprod} class is added.

\code{new_viralprod_class_2} checks the original abundances data frame. If the requirements are met, \code{viralprod_analyze} class is added.
}
\examples{
\dontrun{
# Setup
data_NJ2020_all <- read.csv(system.file('extdata', 
'NJ2020_Station_2_and_6_all_populations.csv', package = "viralprod"))

original_abundances_NJ2020 <- read.csv(system.file('extdata',
'NJ2020_original_abundances.csv', package = "viralprod"))

# Perform
data_S3_class <- new_viralprod_class(data_NJ2020_all)

original_abundances_S3_class <- new_viralprod_class_2(original_abunances_NJ2020)
}
}
