% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{tp}
\alias{tp}
\title{Adding time ranges of assay to dataframe}
\usage{
tp(DF)
}
\arguments{
\item{DF}{Dataframe consisting of a numerical column Timepoint that has the different timepoints}
}
\value{
The input dataframe added with two columns Time_Range and Time_Time that show the different time ranges of the assay
}
\description{
Adding time ranges of assay to dataframe
}
\examples{
x <- data.frame(Timepoint = c(0,3,6,9,12,24))
tp(x)
}
