% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_viral_production.R
\name{visualize_viral_production}
\alias{visualize_viral_production}
\title{Wrapper function to visualize viral production data}
\usage{
visualize_viral_production(
  data = data.frame(),
  original_abundances = data.frame(),
  write_csv = TRUE,
  output_dir = ""
)
}
\arguments{
\item{data}{Data frame with the output of the flow cytometry.}

\item{original_abundances}{Data frame with the abundances of bacterial and virus population in the original sample.}

\item{write_csv}{If \code{TRUE}, the output data frames will be saved as csv files in the same folder of the viral production calculation.
If no csv files are wanted, set to \code{FALSE}. (Default = \code{TRUE})}

\item{output_dir}{String that refers to the location of folder to save the data frames as csv files.}
}
\value{
Plot objects will be stored in variable \code{plot_list} in the global environment. A subfolder, \code{Figures}, will be created in the given output folder for storing the figures as PDFs.
}
\description{
A major step in data analyses is \verb{Data Visualization}. Some suggestions to visualize the flow cytometry, viral
production or analyzed viral production data are given. The different visualizations will be performed, stored
and saved as PDFs if wanted.

More details about the previous performed steps can be found on:
\enumerate{
\item Calculate viral production step: \link{calculate_viral_production}
\item Analyze viral production step: \link{analyze_viral_production}
\item Visualizations of viral production data: \link{vp_visuals}
}
}
\examples{
\dontrun{
# Setup
data_NJ2020_all <- read.csv(system.file('extdata', 
'NJ2020_Station_2_and_6_all_populations.csv', package = "viralprod"))

original_abundances_NJ2020 <- read.csv(system.file('extdata',
'NJ2020_original_abundances.csv', package = "viralprod"))

#' # Perform
# Write output files
visualize_viral_production(data_NJ2020_all, original_abundances_NJ2020, 
output_dir = paste0(system.file(“extdata”, package = “viralprod”), 
“/NJ2020_vp_results”))

# No output files
visualize_viral_production(data_NJ2020_all, original_abundances_NJ2020, write_csv = F)
}
}
