% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_dataframe_functions.R
\name{vp_dataframes}
\alias{vp_dataframes}
\alias{vp_separate_replicate_dataframe}
\alias{vp_average_replicate_dataframe}
\title{Construct viral count data frame}
\usage{
vp_separate_replicate_dataframe(
  data,
  keep_0.22_samples = FALSE,
  add_timepoints = TRUE
)

vp_average_replicate_dataframe(data, add_timepoints = TRUE)
}
\arguments{
\item{data}{Dataframe with the output of the flow cytometer (Step 1).}

\item{keep_0.22_samples}{If \code{FALSE}, 0.22 samples will be removed from data. These represent the control samples and are normally always filtered out. If you want to keep control samples, set to \code{TRUE}. (Default = \code{FALSE})}

\item{add_timepoints}{If \code{TRUE}, different time ranges will be added with \code{vp_add_timepoints()}. If \code{FALSE}, no addition of time ranges to data frame. (Default = \code{TRUE})}
}
\value{
Dataframe with the count for each population for each sample at the different time points of the assay. If \code{add_timepoints = T}, a column with the time ranges is added to the dataframe.
}
\description{
Determining the bacterial and viral counts.

\code{vp_separate_replicate_dataframe} determines the counts by taking the separate replicates into account.

\code{vp_average_replicate_dataframe} determines the counts by taking the average over the replicates.
}
\examples{
\dontrun{
data_NJ2020 <- read.csv(system.file('extdata', 'NJ2020_subset.csv', package = "viralprod"))

vp_separate_replicate_dataframe(data_NJ2020)
vp_separate_replicate_dataframe(data_NJ2020, add_timepoints = F)
vp_separate_replicate_dataframe(data_NJ2020, keep_0.22_samples = T)

vp_average_replicate_dataframe(data_NJ2020)
vp_average_replicate_dataframe(data_NJ2020, add_timepoints = F)
}
}
