% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_list_of_methods.R
\name{vp_list_of_methods}
\alias{vp_list_of_methods}
\title{List of methods used for viral production calculation}
\usage{
vp_list_of_methods()
}
\value{
A list with the different variants of linear regression and VIPCAL used to calculate viral production.
}
\description{
A list of all the different variants of linear regression and VIPCAL used to calculate the viral
production rate.

More details on the different variants:
\itemize{
\item Variants of linear regression: \link{vp_methods_linear}
\item Variants of VIPCAL: \link{vp_methods_VIPCAL}
}
}
