% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_slope_functions.R
\name{determine_vp_LM}
\alias{determine_vp_LM}
\alias{determine_vp_linear_allpoints}
\alias{determine_vp_linear_separate_replicates}
\alias{determine_vp_linear_average_replicates}
\alias{determine_vp_linear_LMER_model}
\title{Calculate viral production with linear regression}
\usage{
determine_vp_linear_allpoints(SR_dataframe)

determine_vp_linear_separate_replicates(SR_dataframe)

determine_vp_linear_average_replicates(AVG_dataframe)

determine_vp_linear_LMER_model(SR_dataframe)
}
\arguments{
\item{SR_dataframe}{Data frame with the viral counts and time ranges, see \link{vp_separate_replicate_dataframe} for more details.}

\item{AVG_dataframe}{Data frame with the viral counts and time ranges, see \link{vp_average_replicate_dataframe} for more details.}
}
\value{
Data frame with the viral production rate and the absolute viral production for each population at given time range of the assay.
}
\description{
\verb{Linear Regression} uses the slope between the viral counts to determine the viral production rate.
Lytic viral production can be derived from the slope of the VP samples. For lysogenic viral production,
the slope of VP samples needs to be subtracted from the slope of VPC samples. See \link[stats:lm]{stats::lm} for more
details on the linear model used to determine the slope.

\code{determine_vp_linear_allpoints} considers all points when calculating the slopes, no distinguishing between replicates.

\code{determine_vp_linear_separate_replicates} uses a separate replicate treatment, distinguishing between replicates.

\code{determine_vp_linear_average_replicates} uses an average replicate treatment, average over the replicates.

\code{determine_vp_linear_LMER_model} uses an average replicate treatment, average over the replicates is included in LMER model.
Difference curve estimation by LMER model, See \link{vp_LMER_model} for more details about LMER model.
}
\examples{
\dontrun{
data_NJ2020 <- read.csv(system.file('extdata', 'NJ2020_subset.csv', package = "viralprod"))
DF_SR <- vp_separate_replicate_dataframe(data_NJ2020)
DF_AVG <- vp_average_replicate_dataframe(data_NJ2020)

determine_vp_linear_allpoints(DF_SR)

determine_vp_linear_separate_replicates(DF_SR)

determine_vp_linear_average_replicates(DF_AVG)

determine_vp_linear_LMER_model(DF_SR)
}
}
