% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_slope_functions.R
\name{determine_vp_linear_LMER_model}
\alias{determine_vp_linear_LMER_model}
\title{Determine viral production by linear regression with difference curve estimation by LMER model}
\usage{
determine_vp_linear_LMER_model(SR_dataframe)
}
\arguments{
\item{SR_dataframe}{Dataframe with the viral counts and time ranges, see \link{vp_separate_replicate_dataframe} for more details.}
}
\value{
Dataframe with the viral production rate and the absolute viral production for each population averaged over the replicates at given time range of the assay. Difference curve estimated by LMER model.
}
\description{
Viral production rate is calculated by using average replicate treatment, average over the replicates is taken
and the difference curve is estimated by the LMER model instead by subtraction. See \link{LMER_model} for
more details about LMER model. Use of linear model to determine the slope between the viral counts on
different time points of the assay. See \link[stats:lm]{stats::lm} for more details on the linear model.
}
\examples{
\dontrun{
data_NJ2020 <- read.csv(system.file('extdata', 'NJ2020_subset.csv', package = "viralprod"))
DF_SR <- vp_separate_replicate_dataframe(data_NJ2020)

viral_production_linear_LMER <- determine_vp_linear_LMER_model(DF_SR)
}
}
