% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_slope_functions.R
\name{determine_vp_linear_average_replicates}
\alias{determine_vp_linear_average_replicates}
\title{Determine viral production by linear regression with average replicate treatment}
\usage{
determine_vp_linear_average_replicates(AVG_dataframe)
}
\arguments{
\item{AVG_dataframe}{Dataframe with the viral counts and time ranges, see \link{vp_average_replicate_dataframe} for more details.}
}
\value{
Dataframe with the viral production rate and the absolute viral production for each population averaged over the replicates at given time range of the assay.
}
\description{
Viral production rate is calculated by using average replicate treatment, average over the replicates
is taken. Use of linear model to determine the slope between the viral counts on different time points of the assay.
See \link[stats:lm]{stats::lm} for more details on the linear model.
}
\examples{
\dontrun{
data_NJ2020 <- read.csv(system.file('extdata', 'NJ2020_subset.csv', package = "viralprod"))
DF_AVG <- vp_average_replicate_dataframe(data_NJ2020)

viral_production_linear_AR <- determine_vp_linear_average_replicates(DF_AVG)
}
}
