% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_methods_linear.R
\name{vp_methods_linear}
\alias{vp_methods_linear}
\alias{vp_linear_allpoints}
\alias{vp_linear_separate_replicates}
\alias{vp_linear_average_replicates}
\alias{vp_linear_average_replicates_diff}
\alias{vp_linear_average_replicates_diff_LMER}
\title{Viral production calculation by linear regression}
\usage{
vp_linear_allpoints(data)

vp_linear_separate_replicates(data, AVG = T)

vp_linear_average_replicates(data)

vp_linear_average_replicates_diff(data)

vp_linear_average_replicates_diff_LMER(data)
}
\arguments{
\item{data}{Dataframe with the output of the flow cytometer (Step 1).}

\item{AVG}{Interested in the lytic and lysogenic viral production. To study the lysogenic viral production,
in need of difference samples. If \code{TRUE}, average over replicates after determining viral production with
separate replicate treatment and lysogenic viral production is calculated. If \code{FALSE}, no averaging over replicates,
output will consits of only VP and VPC samples with separate replicate treatment. (Default = \code{TRUE})}
}
\value{
Dataframe with the viral production rate and the absolute viral production for each population at given time range of the assay.
}
\description{
Two main methods are used to determine the viral production rate over time in our assay: \verb{Linear Regression} vs
\code{VIPCAL}. Different variants of both methods are performed, variance in replicate treatment, use of standard error
or type of estimation of difference curve is presented. Next, the different variants of \verb{Linear Regression},
which uses the slope between the viral counts to determine the viral production rate. The following general
step-by-step plan is carried out by each of the variants:
\enumerate{
\item Constructing correct data frame depending on replicate treatment.
\item Calculate viral production depending on replicate treatment, standard error use and difference curve estimation.
If no difference curve estimation, calculate lysogenic viral production.
\item Arrange output dataframe.
}

\code{vp_linear_allpoints} considers all points: no replicate treatment, no SE taken into account, no difference curve estimation.

\code{vp_linear_separate_replicates}: separate replicate treatment, no SE taken into account, no difference curve estimation.

\code{vp_linear_average_replicates}: average replicate treatment, no SE taken into account, no difference curve estimation.

\code{vp_linear_average_replicates_diff}: average replicate treatment, no SE taken into account, difference curve estimation by subtraction.

\code{vp_linear_average_replicates_diff_LMER}: average replicate treatment, no SE taken into account, difference curve estimation by LMER model.

More details about the used functions:
\itemize{
\item Constructing data frames: \link{vp_dataframes}
\item Calculating viral production with linear regression: \link{determine_vp_LM}
\item LMER model: \link{vp_LMER_model}
\item Calculating lyosgenic production: \link{vp_calculate_difference_samples}
}
}
\examples{
\dontrun{
data_NJ2020 <- read.csv(system.file('extdata', 'NJ2020_subset.csv', package = "viralprod"))

vp_linear_allpoints(data_NJ2020)

vp_linear_separate_replicates(data_NJ2020)
vp_linear_separate_replicates(data_NJ2020, AVG = F)

vp_linear_average_replicates(data_NJ2020)

vp_linear_average_replicates_diff(data_NJ2020)

vp_linear_average_replicates_diff_LMER(data_NJ2020)
}
}
