% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_methods_VIPCAL.R
\name{vp_methods_VIPCAL}
\alias{vp_methods_VIPCAL}
\alias{vp_VIPCAL_separate_replicates}
\alias{vp_VIPCAL_average_replicates}
\alias{vp_VIPCAL_average_replicates_SE}
\alias{vp_VIPCAL_average_replicates_diff}
\alias{vp_VIPCAL_average_replicates_diff_SE}
\alias{vp_VIPCAL_average_replicates_diff_LMER}
\alias{vp_VIPCAL_average_replicates_diff_LMER_SE}
\title{VIPCAL variants for viral production calculation}
\usage{
vp_VIPCAL_separate_replicates(data, AVG = T)

vp_VIPCAL_average_replicates(data)

vp_VIPCAL_average_replicates_SE(data)

vp_VIPCAL_average_replicates_diff(data)

vp_VIPCAL_average_replicates_diff_SE(data)

vp_VIPCAL_average_replicates_diff_LMER(data)

vp_VIPCAL_average_replicates_diff_LMER_SE(data)
}
\arguments{
\item{data}{Data frame with the output of the flow cytometry.}

\item{AVG}{One of the methods uses an separate replicate treatment. In the end, interested in the lytic and lysogenic
viral production. For the last one, difference samples are needed. If \code{TRUE}, average over replicates after determining viral production with
separate replicate treatment and lysogenic viral production is calculated. If \code{FALSE}, no averaging over replicates,
output will consists of only VP and VPC samples with separate replicate treatment. (Default = \code{TRUE})}
}
\value{
Data frame with the viral production rate and the absolute viral production for each population at the different time points of the assay.
}
\description{
Two main methods are used to determine the viral production rate over time in our assay: \verb{Linear Regression} vs
\code{VIPCAL}. Different variants of both methods are performed: variance in replicate treatment, use of standard error
and type of estimation of the difference curve is presented. Next, the different variants of \code{VIPCAL},
which uses the average of increments between the viral counts to determine the viral production rate.
The following general step-by-step plan is carried out by each of the variants:
\enumerate{
\item Constructing correct data frame depending on replicate treatment.
\item Calculate viral production depending on replicate treatment, standard error use and difference curve estimation.
If no difference curve estimation, calculate lysogenic viral production.
\item Arrange output data frame.
}

\code{vp_VIPCAL_separate_replicates}: separate replicate treatment, no SE taken into account, no difference curve estimation.

\code{vp_VIPCAL_average_replicates}: average replicate treatment, no SE taken into account, no difference curve estimation.

\code{vp_VIPCAL_average_replicates_SE}: average replicate treatment, SE taken into account, no difference curve estimation.

\code{vp_VIPCAL_average_replicates_diff}: average replicate treatment, no SE taken into account, difference curve estimation by subtraction.

\code{vp_VIPCAL_average_replicates_diff_SE}: average replicate treatment, SE taken into account, difference curve estimation by subtraction.

\code{vp_VIPCAL_average_replicates_diff_LMER}: average replicate treatment, no SE taken into account, difference curve estimation by LMER model.

\code{vp_VIPCAL_average_replicates_diff_LMER_SE}: average replicate treatment, SE taken into account, difference curve estimation by LMER model.

More details about the used functions:
\itemize{
\item Constructing data frames: \link{vp_dataframes}, \link{vp_check_populations}
\item Calculating viral production with VIPCAL: \link{determine_vp_VIPCAL}
\item LMER model: \link{vp_LMER_model}
\item Calculating lysogenic production: \link{vp_calculate_difference_samples}
}
}
\examples{
\dontrun{
data_NJ2020_all <- read.csv(system.file('extdata', 
'NJ2020_Station_2_and_6_all_populations.csv', package = "viralprod"))
vp_check_populations(data_NJ2020_all)

vp_VIPCAL_separate_replicates(data_NJ2020_all)
vp_VIPCAL_separate_replicates(data_NJ2020_all, AVG = F)

vp_VIPCAL_average_replicates(data_NJ2020_all)

vp_VIPCAL_average_replicates_SE(data_NJ2020_all)

vp_VIPCAL_average_replicates_diff(data_NJ2020_all)

vp_VIPCAL_average_replicates_diff_SE(data_NJ2020_all)

vp_VIPCAL_average_replicates_diff_LMER(data_NJ2020_all)

vp_VIPCAL_average_replicates_diff_LMER_SE(data_NJ2020_all)
}
}
