% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_VIPCAL_functions.R
\name{determine_vp_VIPCAL}
\alias{determine_vp_VIPCAL}
\alias{determine_vp_VIPCAL_separate_replicates}
\alias{determine_vp_VIPCAL_average_replicates}
\alias{determine_vp_VIPCAL_average_replicates_SE}
\alias{determine_vp_VIPCAL_LMER_model}
\alias{determine_vp_VIPCAL_LMER_model_SE}
\title{Calculate viral production with VIPCAL}
\usage{
determine_vp_VIPCAL_separate_replicates(SR_dataframe)

determine_vp_VIPCAL_average_replicates(AVG_dataframe)

determine_vp_VIPCAL_average_replicates_SE(AVG_dataframe)

determine_vp_VIPCAL_LMER_model(SR_dataframe)

determine_vp_VIPCAL_LMER_model_SE(SR_dataframe)
}
\arguments{
\item{SR_dataframe}{Data frame with the viral counts and time ranges, see \link{vp_separate_replicate_dataframe} for more details.}

\item{AVG_dataframe}{Data frame with the viral counts and time ranges, see \link{vp_average_replicate_dataframe} for more details.}
}
\value{
Data frame with the viral production rate and the absolute viral production for each population at given time range of the assay.
}
\description{
On the contrary of Linear Regression, \code{VIPCAL} uses the average of increments to determine
the viral production rate. After determining peaks and valleys in the count data, averaging the increases over
time will result in the viral production rate. Lytic viral production can be derived as the average
of increments of the VP samples. For lysogenic viral production, the average of increments of the difference
curve needs to be looked at. \code{VIPCAL-SE} goes one step further and takes the standard error into account. With
the determination of peaks and valleys, only true increments are returned. A peak/valley is only defined if
there is no overlap between the standard errors. By doing this, sufficient differences in count values,
representing possible increments, are preserved.

\code{determine_vp_VIPCAL_separate_replicates} uses a separate replicate treatment, distinguishing between replicates.

\code{determine_vp_VIPCAL_average_replicates} uses an average replicate treatment, average over the replicates.

\code{determine_vp_VIPCAL_average_replicates_SE} uses VIPCAL-SE and an average replicate treatment, average over the replicates.

\code{determine_vp_VIPCAL_LMER_model} uses an average replicate treatment, average over the replicates is included in LMER model.
Difference curve estimation by LMER model, See \link{vp_LMER_model} for more details about LMER model.

\code{determine_vp_VIPCAL_LMER_model_SE} Uses VIPCAL-SE and an average replicate treatment, average over the replicates is included in LMER model.
Difference curve estimation by LMER model, See \link{vp_LMER_model} for more details about LMER model.
}
\examples{
\dontrun{
data_NJ2020 <- read.csv(system.file('extdata', 'NJ2020_subset.csv', package = "viralprod"))
DF_SR <- vp_separate_replicate_dataframe(data_NJ2020)
DF_AVG <- vp_average_replicate_dataframe(data_NJ2020)

determine_vp_VIPCAL_separate_replicates(DF_SR)

determine_vp_VIPCAL_average_replicates(DF_AVG)

determine_vp_VIPCAL_average_replicates_SE(DF_AVG)

determine_vp_VIPCAL_LMER_model(DF_SR)

determine_vp_VIPCAL_LMER_model_SE(DF_SR)
}
}
