% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_VIPCAL_functions.R
\name{determine_vp_VIPCAL_average_replicates_SE}
\alias{determine_vp_VIPCAL_average_replicates_SE}
\title{Determine viral production by VIPCAL-SE with average replicate treatment}
\usage{
determine_vp_VIPCAL_average_replicates_SE(AVG_dataframe)
}
\arguments{
\item{AVG_dataframe}{Dataframe with the viral counts and time ranges, see \link{vp_average_replicate_dataframe} for more details.}
}
\value{
Dataframe with the viral production rate and the absolute viral production for each population averaged over the replicates at given time range of the assay. Standard error is taken into account.
}
\description{
Viral production rate is calculated by using average replicate treatment, average over the replicates
is taken. VIPCAL-SE determines viral production rate as the average of increments between the viral counts
on different time points of the assay. VIPCAL-SE works the same as VIPCAL but takes the standard error into
account. When determining peaks and valleys, only true increments are kept. A peak/valley is only defined if
there is no overlap between the standard errors. By doing this, sufficient differences in count values,
representing possible increments, are preserved.
}
\examples{
\dontrun{
data_NJ2020 <- read.csv(system.file('extdata', 'NJ2020_subset.csv', package = "viralprod"))
DF_AVG <- vp_average_replicate_dataframe(data_NJ2020)

viral_production_VIPCAL_AR_SE <- determine_vp_VIPCAL_average_replicates_SE(DF_AVG)
}
}
