% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_calculate_DIFF_samples.R
\name{vp_calculate_difference_samples}
\alias{vp_calculate_difference_samples}
\title{Calculate lysogenic viral production}
\usage{
vp_calculate_difference_samples(DF, VIPCAL = FALSE, SE = FALSE)
}
\arguments{
\item{DF}{Data frame containing the viral production for VP and VPC samples calculated with linear regression or VIPCAL.}

\item{VIPCAL}{If \code{FALSE}, viral production is calculated with linear regression. If viral production is calculated with VIPCAL, set to \code{TRUE}. (Default = \code{FALSE})}

\item{SE}{If \code{FALSE}, viral production is calculated with VIPCAL without taking the standard error into account. If VIPCAL takes the SE into account, set to \code{TRUE}. (Default = \code{FALSE})}
}
\value{
Data frame with the viral production added for lysogenic phase. In the column Sample_Type a new sample, \code{Diff} is introduced that represents the lysogenic viral production.
}
\description{
Viral production can be divided in two phases: lytic and lysogenic viral production. The VP samples
cover the lytic viral production, VPC samples include both phases since treatment with the antibiotic \code{mitomycin-C}.
If the bacteriophage is in the lysogenic phase, it integrates with the genome of the bacteria and
can't be measured. Since mitomycin C inhibits DNA synthesis in the bacteria, bacteriophages go into the lytic
phase and measurement is possible. Viral production is calculated by different variants of \verb{Linear Regression}
or \code{VIPCAL}. Some of these variants have the calculation of the lysogenic viral production included by estimating
a difference curve by subtraction or LMER model. For the other variants, calculate lysogenic viral production
afterwards.
}
\examples{
\dontrun{
data_NJ2020 <- read.csv(system.file('extdata', 'NJ2020_subset.csv', package = "viralprod"))
DF_SR <- vp_separate_replicate_dataframe(data_NJ2020)
DF_AVG <- vp_average_replicate_dataframe(data_NJ2020) \%>\% subset(Sample_Type != 'Diff')

vp_linear_allpoints <- determine_vp_linear_allpoints(DF_SR)
vp_VIPCAL_AR <- determine_vp_VIPCAL_average_replicates(DF_AVG)
vp_VIPCAL_AR_SE <- determine_vp_VIPCAL_average_replicates_SE(DF_AVG)

vp_calculate_difference_samples(vp_linear_allpoints)
vp_calculate_difference_samples(vp_VIPCAL_AR, VIPCAL = T)
vp_calculate_difference_samples(vp_VIPCAL_AR_SE, VIPCAL = T, SE = T)
}
}
